requirements.txt
asgiref==3.2.10
coverage==5.3
dj-database-url==0.5.0
Django==3.1.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
mock==4.0.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
